+++ src/index.ts
@@ -1,6 +1,6 @@
 import { atob, btoa } from 'abab';
 import * as check from 'check-types';
-import { defaultsDeep } from 'lodash';
+import defaultsDeep from 'lodash.defaultsdeep';
 import { AchievementIconsResource } from './achievementIcons';
 import { AchievementsResource } from './achievements';
 import { AnalyticsResource } from './analytics';
@@ -9,12 +9,14 @@ import { ApplicationsResource } from './applications';
 import { AwardsResource } from './awards';
 import { CriteriaResource } from './criteria';
 import { EarnedAchievementsResource } from './earnedAchievements';
+import { EarnedAwardsResource } from './earnedAwards';
 import { EventsResource } from './events';
 import { BadgeUpHttp } from './http';
 import { MetricsResource } from './metrics';
 import { ProgressResource } from './progress';
 import { ResourceContext } from './utils/ResourceContext';
 import { VirtualMetricsResource } from './virtualMetrics';
+import { WebhooksResource } from './webhooks';
 
 export class BadgeUp implements ResourceContext {
 
@@ -35,7 +37,9 @@ export class BadgeUp implements ResourceContext {
     public virtualMetrics: VirtualMetricsResource;
     public events: EventsResource;
     public progress: ProgressResource;
+    public earnedAwards: EarnedAwardsResource;
     public achievementIcons: AchievementIconsResource;
+    public webhooks: WebhooksResource;
 
     /**
      * Construct an instance of the BadgeUp client.
@@ -94,6 +98,8 @@ export class BadgeUp implements ResourceContext {
         this.events = new EventsResource(this);
         this.progress = new ProgressResource(this);
         this.achievementIcons = new AchievementIconsResource(this);
+        this.earnedAwards = new EarnedAwardsResource(this);
+        this.webhooks = new WebhooksResource(this);
     }
 }
 
@@ -137,8 +143,10 @@ export * from './applications/Application.class';
 export * from './awards/Award.class';
 export * from './criteria/Criterion.class';
 export * from './earnedAchievements/EarnedAchievement.class';
+export * from './earnedAwards/EarnedAward.class';
 export * from './events/Event.class';
 export * from './metrics/Metric.class';
 export * from './progress/Progress.class';
+export * from './webhooks/Webhook.class';
 export * from './utils/Meta.class';
 export * from './utils/JsonPatch.class';