+++ src/progress/Progress.class.ts
@@ -1,6 +1,7 @@
+import { Award, Criterion } from '..';
 import { Achievement } from '../achievements/Achievement.class';
 
-export interface Progress {
+export interface BaseProgress {
     /**
      * States if the achievement has been earned.
      */
@@ -18,13 +19,21 @@ export interface Progress {
      */
     achievementId: string;
     /**
-     * The root achievement document.
-     */
-    achievement: Achievement;
-    /**
      * A string that uniquely identifies this most recently earned achievement.
      */
     earnedAchievementId: string;
+}
+
+export interface Progress extends BaseProgress {
+    /**
+     * Included achievement and corresponding resource information
+     */
+    achievement?: Achievement & {
+        resources?: {
+            criteria?: Criterion[];
+            awards?: Award[];
+        };
+    };
 }
 
 export interface ProgressTreeGroup {